// src/module/client/client.ts
export interface IClient {
    clientid: number;
    clientcode: string;
    clientname: string;
    contactperson: string;
    salesman: string;
    ymrep: string;
    phone_no: string;
    tel_no: string;
    email: string;
    address: string;
    discount: number;
    moneytype: string;
    useflag: number;
    level1review: number;
    level1name: string;
    level1date: Date | null;
    level2review: number;
    level2name: string;
    level2date: Date | null;
    remark1: string;
    remark2: string;
    remark3: string;
    remark4: string;
    remark5: string;
    remark6: string;
    remark7: string;
    remark8: string;
    remark9: string;
    remark10: string;
    creater: string;
    createdAt: Date | null;
    updater: string;
    updatedAt: Date | null;
    clientareaid: number | null;
    operateareaid: number | null;
    currencyid: number | null;
    del_uuid: number;
    deletedAt: Date | null;
    deleter: string;
    gs:number;
}

export class Client implements IClient {
    clientid: number = 0;
    clientcode: string = "";
    clientname: string = "";
    contactperson: string = "";
    salesman: string = "";
    ymrep: string = "";
    phone_no: string = "";
    tel_no: string = "";
    email: string = "";
    address: string = "";
    discount: number = 0;
    moneytype: string = "";
    useflag: number = 1;
    level1review: number = 0;
    level1name: string = "";
    level1date: Date | null = null;
    level2review: number = 0;
    level2name: string = "";
    level2date: Date | null = null;
    remark1: string = "";
    remark2: string = "";
    remark3: string = "";
    remark4: string = "";
    remark5: string = "";
    remark6: string = "";
    remark7: string = "";
    remark8: string = "";
    remark9: string = "";
    remark10: string = "";
    creater: string = "";
    createdAt: Date | null = null;
    updater: string = "";
    updatedAt: Date | null = null;
    clientareaid: number | null = null;
    operateareaid: number | null = 0;
    currencyid: number | null = null;
    del_uuid: number = 0;
    deletedAt: Date | null = null;
    deleter: string = "";
    gs:number = 0;
}